Graph Constructor from file
Changed Graph class. Major change: added constructor for reading a graph from file. "mst_data.txt" is such a graph file in the correct format for this new constructor. Also changed distance type from float to int. This was mostly done for debugging, since I suspected an approximation error was the cause of a bug: it wasn't. Decided to leave it with int anyways. The best thing to do would be for it to be generic and accept either ints of floats, but I'll leave that idea for another time.
Changes to MinimumSpanningTree: Mostry changes to main function to properly use the new constructor from Graph class. Also added a commentary, explaining the occasion of writing this class, as well as a new method to get the total cost of the tree, since the coursera exercise asked for this.